requirements.txt
fastapi
uvicorn[standard]
cryptography
python-dotenv
